repo/backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Data validation
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite>=0.19.0

# Auth / security
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0
cryptography==42.0.7

# File handling / watermarking
pillow==10.3.0
reportlab==4.2.0
pypdf==4.2.0

# Structured logging
structlog==24.2.0

# HTTP client (for internal service calls and tests)
httpx==0.27.0

# Background tasks
apscheduler==3.10.4

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
anyio==4.3.0

# Utilities
python-multipart==0.0.9
